Fix bugs and rename files
