+++ examples/conversation_basic/conversation.py
@@ -2,7 +2,7 @@
 """A simple example for conversation between user and assistant agent."""
 import agentscope
 from agentscope.agents import DialogAgent
-from agentscope.agents.user_agent import UserAgent
+from agentscope.agents._user_agent import UserAgent
 from agentscope.pipelines.functional import sequentialpipeline
