+++ examples/distributed_parallel_optimization/main.py
@@ -7,7 +7,7 @@ from searcher_agent import SearcherAgent
 from answerer_agent import AnswererAgent
 
 import agentscope
-from agentscope.agents.user_agent import UserAgent
+from agentscope.agents._user_agent import UserAgent
 from agentscope.message import Msg
