+++ src/agentscope/formatters/__init__.py
@@ -1,10 +1,10 @@
 # -*- coding: utf-8 -*-
 """The formatter modules for different models."""
-from .formatter_base import FormatterBase
-from .common_formatter import CommonFormatter
-from .anthropic_formatter import AnthropicFormatter
-from .gemini_formatter import GeminiFormatter
-from .openai_formatter import OpenAIFormatter
+from ._formatter_base import FormatterBase
+from ._common_formatter import CommonFormatter
+from ._anthropic_formatter import AnthropicFormatter
+from ._gemini_formatter import GeminiFormatter
+from ._openai_formatter import OpenAIFormatter
 
 __all__ = [
     "FormatterBase",
