+++ src/agentscope/formatters/_common_formatter.py
