+++ src/agentscope/formatters/_formatter_base.py
