+++ src/agentscope/models/dashscope_model.py
@@ -8,7 +8,7 @@ from typing import Any, Union, List, Optional, Generator
 from loguru import logger
 
 from ..formatters import CommonFormatter
-from ..formatters.dashscope_formatter import DashScopeFormatter
+from ..formatters._dashscope_formatter import DashScopeFormatter
 from ..manager import FileManager
 from ..message import Msg, ToolUseBlock
